Add Writer#display_banner test
Note: The first calculator test cannot be completed
until we have the Writer and Reader helpers, so we
start with the Writer.

The first task the Writer needs to do is to print the
welcome banner. The stream that the writer will write
to is passed as a parameter, which allows us to swap
in a StringIO to do the testing.