Add Reader to Calculator to read input
Previous commit was to test the Reader#read method,
now the read call is added to Calculator#compute and
the Reader is made the default value for the first
arg of compute. In the CalculatorTest, the MockReader
needed this method added. The CalculatorTest uses an
array to hold the fake responses from the user in an
array, so the mock read simply shifts an element off
this array.