Add ask_for_number to Calculator#compute
Modified version of ask_for_number added to the
compute method. Note that through all this, the
first test we wrote against the compute method
is still failing, and has since it was written.
This is how to "bootstrap" an integration type
test -- get the Reader and Writer working enough
to pass the first happy-path test on compute,
then test any further functionality on compute.

Note, this commit was created with git add --patch,
which allows you to select the diffed lines from a
file to add to the commit. I did this because I
forgot to finish the ask_for_number commit before
plunging ahead.

The patch idea was an experiment, so we'll see how
it turns out.