Calculator#compute complete to first video
This is a big diff, all the rest of the integration
style tests are here, along with a helper method
Calculator#calculate. Here is a synopsis of each
method that was added or modified:

* Calculator#compute
The rest of the if-else statement was fleshed out.
Mixed in is a refactoring of the way the result is
computed and a refactoring of the way compute returns
values. To prevent the noise of duplicating all the
to_i and to_f calls, I created the helper calculate
that takes the arithmetic operator as an argument. The
return from calculate is fed into display_result, which
seems to read nice. The method Writer#display_result was
modified to return the computed value after it displays.
Since the result is the desired return value, we can
return immediately after display returns in each
branch of the if-else.

* Calculator#calculate
Converts string arguments representing integers into
numeric types and does the computation requested. Returns
the computed result.

* WriterTest#display_result
Modified to test that the result passed to display_result
is returned. Previously, the display was produced with no
return value (it was nil from output_stream.puts call).

* CalculatorTest#<xxx>

These tests are integration style tests, they use mocks
for the reading and writing, this is to be able to
concentrate on the overall behavior and to make sure
all the parts are bound together well. Each test method
name hopefully explains its provenance or existence.

test_subtract
test_product
test_quotient
test_noop
test_calculate

MockWriter#display_result
Got stuck here, the first CalculatorTest was returning
nil instead of the result and I couldn't understand why.
The mock had all empty methods, so in hindsight it made
perfect sense. Needed to have this method simply pass
the value through.