Add Writer#decorate for pretty prompt
This satisfies the first requirement in the
Refactoring Calculator video. The decorate method
is class-level and private, any message going
to output is hit with this method, which adds a
prompt of => to the message. I wrote a long comment
about how decorate was made class private, see that
for further info.

The design here reflects two ideas: have Writer do the
prompt decorating, not the main calculator method; second,
follow the video as closely as possible, treat it as the
spec. The video simply changed the appearance, there was no
toggle, ability to change to another prompt, etc. This
step is a logical one toward those more complex designs.