Add Reader#valid? method
Previously, the read functions would return whatever
was input (after a chomp). In order to prompt until
a valid number is input, we need this method to return
falsy for non-numbers.

I did the same semi-hack on Reader as I did on Writer,
which is to make valid? a private class method. This is
a helper method local to Reader, no other object has
any business using this function.

Preview: The next commit will add the "prompt until
valid" control flow, it may be useful to squash that
commit with this one.