Add prompt until valid number input logic
Used a loop in calculator.rb to either read a number or print
error message and ask again.

- Writer#display_invalid_number gives error when input is invalid.
Corresponding test is in WriterTest (just checks/specs the output).

- Added tests to ReaderTest for checking zero and nil. These were
part of flailing around when the mock reader was not being
cooperative, but I think they stand on there own, at this point we
are triggering something to happen when read_num returns false so
it is good to confirm the falsy values.

- CalculatorTest adds tests to checks one bad input on first
number prompt, and to check multiple non-numbers to first prompt.

- MockReader adds a counter to monitor how many times read is called.
Also adds an invalid number check to read_num. This caused the most
delays/issues with this commit, I seem to forget that in
CalculatorTest the Mocks are the readers and writers instead
of the real ones. There was no validity check in read_num in the
mock so it never skipped the bad input. Even though the program
ran correctly, so that the problem couldn't be with the real reader,
I kept theorizing that it was. I also didn't want to add logic
to the fake reader, it seems to defeat the purpose, so that perhaps
drove my aversion to looking there. Once I understood the root of
the problem, I considered adding stubbed in values in MockReader
to get the tests to pass, this is seemingly what boxed up mock
libraries like Minitest::Mock have you do, but it seemed just as
well to add the more abstract check (the equivalent of the real check)
and be done with it.

The MockReader issue above seems to point to how the design
and the ways you test the design have a relationship to each
other. Another note I made was that I have started to change
tests that already exist and whether this was an ideal state
of affairs or the price of doing business. I get the sense that
a better design might help avoid this, more to come on the
subject.