Add calculate another option
Another big commit. Loop is added to calculator method to
perform multiple calculations. Because multiple calculations
are possible now, the return value was changed to return an
array of all the calculations performed. (Future feature may
be to include the terms as well).

To see the overall flow to the calculation, which is probably
of most interest, look at the CalculatorTest diffs. The other
tests are to setup the boilerplate of the reader and writer for
prompting and reading, respectively.

The mock reader is a simple shift off the response array. I was
expecting the worst after the validate number issue.