Input name feature; refactor tests
Application now asks for name and reprompts until a
non-empty reply is given. By file, here is a breakdown
of the changes:

calculator.rb
- Add until loop for getting name
- display error if name invalid
- display greeting

calculator_test.rb
- add test name to beginning of all response arrays
- add continue? input to end of all response arrays
- add newlines (\n) to each entry in all responses. A helper
function add_newline was added
- test valid and invalid name entry

writer.rb
- display_name_error
- display_greeting
- writer_test.rb and mock_writer.rb have corresponding
tests and mocks
- not much to discuss -- simple output functions

mock_reader.rb
- have to chomp all responses now; this is less a mock
and more the real thing
- read_name does validation on mock responses

This was a challenge, I had to learn Minitest a little better,
when you add the name prompter, it breaks all the other tests,
which is a bit jarring. I actually restarted twice, I knew I
knew what to do (lol) but it wa a bit hairy and seemed easier
to start over with what I learned from the first attepmt(s).

As far as learning the tools better, I learned how to run
a single test with rake, I needed this because with so many tests
breaking it was hard to tell if I was making any progress as I
incrementally (per test) fixed things. I also started using
binding.pry to debug in combination with running one test
at a time, since debugging in calculator will hit the breakpoint
for every test unless you isolate, and you lose track of
what is happening. Some of this may be caused by a naive
understanding of vim and tmux as well, though I do like my
workflow and efficiency with these tools in general.