Delete redundant tests
Over the course of adding features, some equivalent tests
were created. I was about to start testing the operator
validation test and was going to write a test_operator_happy
kind of test, which is already done. This prompted me
to look for other times I may have done this.

Deleted
- test_no_new_calc
- test_enter_name