Calculator bonus #2: decimal input
As a first overall impression, it was intersting as I staged
the commit that I need to do nothing to the driver function
calculaor.rb. I don't know if that implies anything, just
interesting at this point.

Checking if float is actually an int in hiding is pretty
stupid; adding 1.0 + 1.0 probably shouldn't be Integer 2.
I wanted to cover all the bases of two types being
combined in the operation, and this just sort of popped out
as something fun to try. I do like the use of the lambda
to hide helper functions, since there is no need to expose
convert to the user of calculator.

The tests basically speak for themselves, they exercise
reading and calculating with floats.