Canary test fails
Getting the first test to fail.  The mortgage.rb file has a
mortgage function that does all the work of reading, writing and
computing.  There will be helper classes Reader and
Writer that are injected into mortgage, this allows the helpers
to be tested independently and to relieve the input output
responsibility from the mortgage function.  Injecting these
classes also allows mocks to be used when we test the overall
behavior (like integration tests), and this is what is set up
here.

See this blog post for the details, I'm following pretty
closely what is done there.
    [Testing code that's hard to test](https://8thlight.com/blog/josh-cheek/2011/10/01/testing-code-thats-hard-to-test.html)